Switch logic to handler
